+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## Version 0.3.0
 
 - Complete rewrite of the library
+- Add 'be_failed' and 'be_successful' operation matchers